refactor(canary-web-components): remove duplicate 'x' svg and refactor component code
remove duplicate 'x' svg and refactor component code following this change

4064
