+++ packages/canary-web-components/src/components/ic-time-input/ic-time-input.tsx
@@ -31,7 +31,7 @@ import {
   renderHiddenInput,
   parseTimeHelper,
 } from "../../utils/helpers";
-import Clear from "../../assets/clear-icon.svg";
+import Clear from "../../assets/close-icon.svg";
 import Clock from "../../assets/clock.svg";
 
 import "../../../../web-components/src/components/ic-input-container/ic-input-container";
